chore: Renamed internal components to work with docgen STC-536
