+++ src/components/EditableField/Select/Select.stories.tsx
@@ -1,4 +1,6 @@
-import EditableFieldSelect, { EditableFieldSelectProps } from "./Select"
+import EditableFieldSelect, {
+  EditableFieldSelectProps,
+} from "./EditableFieldSelect"
 import { Meta, Story } from "@storybook/react/types-6-0"
 import { useState } from "react"
 import { action } from "@storybook/addon-actions"
